chore: update all.nested-setup-function.spec.ts
