+++ tests/spec/test-data/nested-setup-function/c.ts
@@ -1,0 +1,5 @@
+import { LogService, BDep } from '@angular/core';
+
+export class C {
+    constructor(private bDep: BDep, private logger: LogService) {}
+}
